[#6][FEAT] 유저 로그인 로직 추가

+++ src/main/java/com/kancho/byeolbyeol/constellation/domain/ConstellationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ConstellationRepository extends JpaRepository<Constellation, Integer> {
+public interface ConstellationRepository extends JpaRepository<Constellation, Long> {
     Optional<Constellation> findByName(String constellation);
 }
